Change base test class to a mixin
Further reorder tests, making the base test class `EndpointTestsMixin`
the base for all endpoint tests, no matter the server being tested.
`TestClient`s are now retrieved through utility functions.
