+++ tests/server/routers/test_links.py
@@ -3,9 +3,18 @@ import unittest
 
 from optimade.models import LinksResponse
 
-from ..test_server import EndpointTests
+from ..utils import EndpointTestsMixin, get_regular_client, get_index_client
 
 
-class LinksEndpointTests(EndpointTests, unittest.TestCase):
+class LinksEndpointTests(EndpointTestsMixin, unittest.TestCase):
+
+    client = get_regular_client()
+    request_str = "/links"
+    response_cls = LinksResponse
+
+
+class IndexLinksEndpointTests(EndpointTestsMixin, unittest.TestCase):
+
+    client = get_index_client()
     request_str = "/links"
     response_cls = LinksResponse
