+++ tests/server/routers/test_references.py
@@ -3,15 +3,19 @@ import unittest
 
 from optimade.models import ReferenceResponseMany, ReferenceResponseOne
 
-from ..test_server import EndpointTests
+from ..utils import EndpointTestsMixin, get_regular_client
 
 
-class ReferencesEndpointTests(EndpointTests, unittest.TestCase):
+class ReferencesEndpointTests(EndpointTestsMixin, unittest.TestCase):
+
+    client = get_regular_client()
     request_str = "/references"
     response_cls = ReferenceResponseMany
 
 
-class SingleReferenceEndpointTests(EndpointTests, unittest.TestCase):
+class SingleReferenceEndpointTests(EndpointTestsMixin, unittest.TestCase):
+
+    client = get_regular_client()
     test_id = "dijkstra1968"
     request_str = f"/references/{test_id}"
     response_cls = ReferenceResponseOne
